Update index.html V12
transactionType: Set to "ISSUANCE" to indicate that the data should be saved in the OUT sheet.

wwrNo: The WWR No. for the OUT sheet.

iirNo: The IIR No. for the OUT sheet.

docsRefNoIssuance: The DOC. REF. NO. for the OUT sheet.

Other fields (date, itemDescription, containerNo, qtyCs, qtyKg, productionDate, timestamp) remain the same as for the IN sheet.